Adding a redirect

+++ database/migrations/2020_06_17_133222_create_teachers_table.php
@@ -17,7 +17,7 @@ class CreateTeachersTable extends Migration
             $table->id();
             $table->string('name');
             $table->string('surname');
-            $table->int('age');
+            $table->tinyInteger('age');
             $table->string('cv_link');
             $table->timestamps();
         });
